SDK: extract tool message formatting (oh-tab-g9ys)
